Move shared folder to shared-webapp utils and infrastructure packages
