+++ application/account-management/WebApp/package.json
@@ -20,6 +20,7 @@
   "devDependencies": {
     "@repo/build": "*",
     "@repo/config": "*",
+    "@repo/infrastructure": "*",
     "@repo/ui": "*"
   },
   "postcss": {
