+++ application/account-management/WebApp/routes/register/-components/actions.ts
@@ -1,6 +1,6 @@
 import { i18n } from "@lingui/core";
 import { z } from "zod";
-import { getApiError, getFieldErrors } from "@/shared/apiErrorListSchema";
+import { getApiError, getFieldErrors } from "@repo/infrastructure/api/ErrorList";
 import { accountManagementApi } from "@/lib/api/client";
 
 const VALIDATION_LIFETIME = 1000 * 60 * 5; // 5 minutes
