--- application/back-office/WebApp/shared/getCamelCase.ts
@@ -1,3 +1,0 @@
-export function getCamelCase(text: string): string {
-  return text[0].toLowerCase() + text.slice(1);
-}
