+++ application/shared-webapp/infrastructure/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@repo/infrastructure",
+  "private": true,
+  "version": "0.0.0",
+  "license": "MIT",
+  "exports": {
+    ".": "./dist/index.js",
+    "./*": "./dist/*.js",
+    "./package.json": "./package.json"
+  },
+  "scripts": {
+    "dev": "tsc -p . -w",
+    "build": "rm -rf ./dist && tsc -p .",
+    "check": "biome check",
+    "format": "biome format"
+  },
+  "devDependencies": {
+    "@repo/config": "*",
+    "@repo/utils": "*"
+  }
+}
