+++ application/shared-webapp/infrastructure/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "display": "PlatformPlatform build configuration",
+  "extends": "@repo/config/typescript/node-library.json",
+  "compilerOptions": {
+    "rootDir": ".",
+    "outDir": "dist"
+  },
+  "exclude": ["node_modules", "dist"]
+}
