+++ application/shared-webapp/utils/string/getCamelCase.ts
